crud routes finished. React and redux added with initial code. CSS

+++ server/controller/goalController.js
@@ -1,12 +1,13 @@
 const expressHandler = require("express-async-handler");
 
 const Goal = require("../model/goalModel");
+const User = require("../model/userModel");
 
 // description: get goals
 // method: GET /api/goals
 // access: Private
 const getGoals = expressHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id });
 
   res.json(goals);
 });
@@ -20,7 +21,10 @@ const createGoal = expressHandler(async (req, res, next) => {
     throw new Error("Text is required");
   }
 
-  const goal = await Goal.create({ text: req.body.text });
+  const goal = await Goal.create({
+    text: req.body.text,
+    user: req.user.id,
+  });
 
   res.json(goal);
 });
@@ -34,6 +38,20 @@ const updateGoal = expressHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error("Goal not found");
+  }
+
+  const user = await User.findById(req.user.id);
+
+  //check for user
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  //check if user is owner of the goal
+  if (goal.user.toString() !== req.user.id) {
+    res.status(401);
+    throw new Error("User not authorised");
   }
 
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -52,6 +70,20 @@ const deleteGoal = expressHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error("Goal not found");
+  }
+
+  const user = await User.findById(req.user.id);
+
+  //check for user
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  //check if user is owner of the goal
+  if (goal.user.toString() !== req.user.id) {
+    res.status(401);
+    throw new Error("User not authorised");
   }
 
   await goal.deleteOne();
